+++ lib/cuke_modeler/adapters/gherkin_30_adapter.rb
@@ -1,0 +1,13 @@
+require_relative 'gherkin_20_adapter'
+
+
+module CukeModeler
+
+  # @api private
+  #
+  # An adapter that can convert the output of version 30.x of the *cucumber-gherkin* gem into input that is consumable
+  # by this gem. Internal helper class.
+  class Gherkin30Adapter < Gherkin20Adapter; end
+
+  private_constant :Gherkin30Adapter
+end